requirements.txt
Django==5.0.3
psycopg2-binary==2.9.9
djangorestframework==3.14.0
Markdown==3.6